refactor: extract chargen menu nodes
