+++ utils/menus/__init__.py
@@ -1,0 +1,1 @@
+# package marker for menus
